+++ WindowsFormsApplication1/settingsWindow.Designer.cs
@@ -104,7 +104,7 @@
             // 
             // buttonCancel
             // 
-            this.buttonCancel.Location = new System.Drawing.Point(336, 108);
+            this.buttonCancel.Location = new System.Drawing.Point(337, 107);
             this.buttonCancel.Name = "buttonCancel";
             this.buttonCancel.Size = new System.Drawing.Size(75, 23);
             this.buttonCancel.TabIndex = 68;