+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - No recent changes.
 
+## 0.0.6 - 2016-12-06
+### Added
+- Toolbar to Archive creator - settings button is functional
+- changed Storm report generator to set the text file location
+- Changed output text file slightly
+
 ## 0.0.5 - 2016-12-06
 ### Added
 - Settings window now has a functionless GUI